Don't use Int where we can use ShoppingListId or ItemId
Really we should be defining these id types to be:

type ShoppingListId
    = ShoppingListId Int

This is preferable because it prevents us from using ShoppingListId in
place of ItemId. Our type aliases don't, because they both specialise
to Int which means that values constrained to these types are
interchangeable. I'll do that shortly in another commit.