Elm client refactor: make Action & Event first-class types
Previously we were conflating the concept of Msg (Elm Architecture
msgs, triggered by things happening in the UI or websocket connection)
with the concepts of Actions (something we want to do on the server)
and Events (something that has happened on the server). These are
domain concepts and so should be separate from Msg and carry their own
data.

This refactor also includes clearer naming for the Msg types, such
that it's obvious that a Msg indicated that something has happened in
the UI or websocket connection.

We're also now expecting to receive a list of ShoppingList from the
server with most events.