Refactor Elm client - split out Event, Action, Config & UpdateHelpers
This purpose of this refactor is to make ShopShare.elm as easy as
possible to understand. That is, I'm trying to optimise for
understanding the app at the highest level possible.

It should now be clear what Msg, Action and Event types are.

We're also starting to expect that we'll have unified UpdateList and
UpdateItem actions, with semantics not unlike that of HTTP PUT.